pyproject.toml
[project]
name = "ecommerce-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-crispy-forms>=2.4",
    "django-ninja>=1.4.5",
    "pyjwt>=2.10.1",
]
